+++ NEWS.md
@@ -1,7 +1,13 @@
 # metR 0.2.9000 
 
 
+## New features
+
 - `Anomaly()` has a new `baseline` argument. 
+
+##  Bugfixes 
+
+- `ReadNetCDF()` now accepts dates as elements for `subset`.
 
 
 # metR 0.2.0